backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic-settings==2.7.0
supabase==2.10.0
mistralai==1.5.0
playwright==1.49.1
python-dotenv==1.0.1
